large-image-viewer/large-image-viewer/requirements.txt
Pillow
PyQt5
numpy
pyinstaller
opencv-python
matplotlib